some int <-> size_t warnings

+++ Audio.Asio.Unmanaged/Looper.cpp
@@ -32,7 +32,8 @@ Looper* Looper::Create(const ILooperConfig& config)
 
 		int sampleCount = config.SampleCount == 0 ? AsioCore::UsePreferredSize : config.SampleCount;
 
-		device->CreateBuffers(config.InputChannelList, config.InputChannelCount, config.OutputChannelList, static_cast<int>(config.OutputChannelCount), sampleCount, config.OutputSaturation);
+		device->CreateBuffers(config.InputChannelList, static_cast<int>(config.InputChannelCount), 
+			config.OutputChannelList, static_cast<int>(config.OutputChannelCount), sampleCount, config.OutputSaturation);
 
 		Looper* pLooper = new Looper(device);
 		pLooper->AddRef();
